fix index and delete messageController

+++ app/Http/Controllers/MessageController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Apartment;
 use App\Models\Message;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class MessageController extends Controller
 {
@@ -14,7 +15,9 @@ class MessageController extends Controller
     public function index()
     {
         // Recupera tutti i messaggi con paginazione
-        $messages = Message::orderBy('created_at', 'desc')->paginate(5);
+        $apartments = Apartment::whereUserId(Auth::id())->get();
+        $apartment_ids = $apartments->pluck('id');
+        $messages = Message::whereIn('apartment_id', $apartment_ids)->orderBy('created_at', 'desc')->paginate(5);
 
         // Passa i dati alla vista
         return view('admin.messages.index', compact('messages'));
@@ -43,8 +46,11 @@ class MessageController extends Controller
     {
         // Trova il messaggio con l'ID specificato
         $message = Message::findOrFail($id);
-
-        // Passa il messaggio alla vista
+        $apartment_id = $message->apartment_id;
+        $apartment = Apartment::findOrFail($apartment_id);
+        if ($apartment->user_id !== Auth::id()) {
+            return to_route('admin.home')->with('type', 'danger')->with('message', 'Non sei autorizzato ad eseguire questa azione');
+        }
         return view('admin.messages.show', compact('message'));
     }
 
@@ -71,6 +77,11 @@ class MessageController extends Controller
     {
         // Trova il messaggio con l'ID specificato e lo elimina
         $message = Message::findOrFail($id);
+        $apartment_id = $message->apartment_id;
+        $apartment = Apartment::findOrFail($apartment_id);
+        if ($apartment->user_id !== Auth::id()) {
+            return to_route('admin.home')->with('type', 'danger')->with('message', 'Non sei autorizzato ad eseguire questa azione');
+        }
         $message->delete();
 
         // Redirect alla lista dei messaggi con un messaggio di successo
